requirements.txt
langchain
langchain-core
langchain-community
langgraph

# Only Google LLM
langchain-google-genai

# RAG / embeddings
faiss-cpu
tiktoken

# Backend
fastapi
uvicorn
requests

# Utils
python-dotenv
langchain-huggingface
langchain_huggingface
langchain langchain-huggingface sentence_transformers